Build Groups of Views
Views define a piece of your UI. They’re the building blocks of your app. You build a complex view by composing it of small, simple views. In this section, you’ll build the header of the timer screen to display the elapsed time and remaining time of a meeting.