Create the Card View
CardView will summarize the DailyScrum model data and display the title, the number of participants, and the duration. You’ll compose the CardView from smaller views, each of which displays a piece of data in the DailyScrum structure.

You’ll update the structure CardView_Previews so that you get immediate visual feedback as you develop the view.