Make the Card View Accessible
Now that CardView is laid out visually, you’ll make it accessible to all users. You’ll add accessibility modifiers to your views to make understanding and navigating your interface in VoiceOver easier.

CardView contains a label with an image of a clock and a number that represents meeting length. It also contains an image of people and a number of attendees.

You’ll add accessibility labels that use string interpolation so that the labels become meaningful descriptions of the view. VoiceOver will read, “4 attendees” and “10 minute meeting.”