Define the Profile Editor
The user profile editor consists primarily of different controls that change individual details in the profile. Some items in the profile, like the badges, aren’t user-editable, so they don’t appear in the editor.

For consistency with the profile summary, you’ll add the profile details in the same order in the editor.