Create a Custom Notification Interface
Your version of Landmarks for watchOS is almost complete. In this final section, you’ll create a notification interface that displays landmark information whenever you receive a notification indicating that you are close to one of your favorite locations.

This section only covers how to display the notification after you receive it. It doesn’t describe how to set up or send notifications.