Add a macOS Target to the Project
Start by adding a macOS target to the project. Xcode adds a new group and set of starter files for the macOS app, along with the scheme needed to build and run the app. You’ll then add some existing files to the new target.

To be able to preview and run the app, be sure your Mac is running macOS Monterey or later.