Add a Built-in Menu Command
When you create an app using the SwiftUI life cycle, the system automatically creates a menu with commonly needed items, like those for closing the front-most window or for quitting the app. SwiftUI lets you add other common commands with built-in behavior, as well as completely custom commands.

In this section, you’ll add a system-provided command that lets the user toggle the sidebar, to be able to get it back after dragging it closed.