Add Preferences with a Settings Scene
Users expect to be able to adjust settings for a macOS app using the standard Preferences menu item. You’ll add preferences to MacLandmarks by adding a Settings scene. The scene’s views define the contents of the preferences window, which you’ll use to control the initial zoom level of the MapView. You communicate the value to the map view, and store it persistently, by using the @AppStorage property wrapper.